Move aircraft_name function to ceasiompyfunctions
